Introduce tests for DateTimeRestrictions.match()
